不用加减乘除做加法

+++ java/剑指offer/T65.java
@@ -7,10 +7,10 @@ package 剑指offer;
  */
 public class T65 {
     public int add(int a, int b) {
-        while(b!=0){
-            int tmp=a^b;
-            b=((int)(a&b)<<1);
-            a=tmp;
+        while (b != 0) {
+            int tmp = a ^ b;
+            b = ((a & b) << 1);
+            a = tmp;
         }
         return a;
     }
